requirements.txt
matplotlib==3.1.3
numpy==1.18.1
jupyter==1.0.0
opencv-python==4.2.0.32
Pillow==2.2.2
